backend/pyproject.toml
[tool.poetry]
name = "connectionhub"
version = "0.1.0"
description = ""
authors = ["Shahsad KP <shahsadkpklr@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
Pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-storages = {extras = ["s3"], version = "^1.14.2"}
boto3 = "^1.34.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
